some fixes and changes in jobs

+++ BackEnd/database/migrations/2024_05_14_112112_create_freelancing_job_competetors_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->foreignId('job_id')->constrained('freelancing_jobs')->cascadeOnDelete();
             $table->text("description");
-            $table->float("salary");
+            $table->double("salary");
             $table->timestamps();
         });
     }
